% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_airpass.R
\docType{data}
\name{airpass}
\alias{airpass}
\title{airpass}
\format{
airpass est une série temporelle.
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
airpass
}
\description{
Le fichier airpass est directement issu du package dataset (AirPassengers)
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
