% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_chomage_bit.R
\docType{data}
\name{chomage_bit}
\alias{chomage_bit}
\title{chomage_bit}
\format{
chomage_bit est une série temporelle.
}
\source{
\url{https://www.insee.fr/} {Site de l'Insee}
}
\usage{
chomage_bit
}
\description{
Le fichier chomage_bit est la serie temporelle du taux de chomage trimestriel brut au sens du BIT
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
