% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cvs_decompo_graph.R
\name{cvs_decompo_graph}
\alias{cvs_decompo_graph}
\title{cvs_decompo_graph}
\usage{
cvs_decompo_graph(x, ...)
}
\arguments{
\item{x}{une série temporelle}

\item{...}{arguments hérités de la fonction decompose}
}
\value{
un graphique plot
}
\description{
Graphique de décomposition d'une série temporelle
}
\examples{
cvs_decompo_graph(airpass,type="multiplicative")
}
