% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_apur_m2.R
\name{data_apur_m2}
\alias{data_apur_m2}
\title{data_apur_m2}
\usage{
data_apur_m2(input_data, var_apur)
}
\arguments{
\item{input_data}{un data fram}

\item{var_apur}{une variable quantitative}
}
\value{
a clean data frame
}
\description{
A fct function that cleans database from extreme values
}
