% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_departements.R
\docType{data}
\name{departements}
\alias{departements}
\title{departements}
\format{
departements est un data.frame constitué de 100 départements (rows) et de 15 variables (col).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
departements
}
\description{
Le fichier departements est constitué des 100 départements français et donne de l'information
économique, démographique et sociale
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
