% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_departements_regr.R
\docType{data}
\name{departements_regr}
\alias{departements_regr}
\title{departements_regr}
\format{
departements est un data.frame constitué de 100 départements (rows) et de 8 variables (col).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
departements_regr
}
\description{
Le fichier departements_regr est constitué des 100 départements français et donne de l'information
socio démo et des indicateurs d'insécurité
}
\author{
Alexis \email{alexis.montaut@insee.fr}
}
\keyword{data}
