% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dygraph_ts.R
\name{dygraph_ts}
\alias{dygraph_ts}
\title{dygraph_ts}
\usage{
dygraph_ts(serie_ts)
}
\arguments{
\item{serie_ts}{une série temporelle}
}
\value{
un graphique dygraph
}
\description{
Chronogramme d'une série temporelle avec dygraph
}
\examples{
dygraph_ts(airpass)
}
