% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_foot.R
\docType{data}
\name{foot}
\alias{foot}
\title{foot}
\format{
foot est un data constitué de 213 lignes (rows) et de 9 variables (cols).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
foot
}
\description{
le fichier foot est issu du site kaggle
Il contient les résultats de 213 équipes lors des coupes du monde de football masculin de 1990 à 2022 (qualification et phases finales).
}
\author{
Alexis \email{alexis.montaut@insee.fr}
}
\keyword{data}
