% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_box.R
\name{format_box}
\alias{format_box}
\title{format_box}
\usage{
format_box(num, ...)
}
\arguments{
\item{num}{une numérique}

\item{...}{des paramètres issus de format}
}
\value{
un character
}
\description{
Fonction qui met en forme une numérique pour rentrer dans une box
de shinidashboard
}
\examples{
moy <- mean(grandile$AGE)
format_box(moy)

}
