% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{gapminder}
\format{
gapminder est un data constitué de 168 (rows) et de 10 variables (col).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
gapminder
}
\description{
The gapminder dataset included into package asta is an exctract from
the database gapminder if the gapminder package
}
\author{
Berangere \email{berangere.duplouy@insee.fr}
}
\keyword{data}
