% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_grandile_apur.R
\docType{data}
\name{grandile_apur}
\alias{grandile_apur}
\title{grandile_apur}
\format{
grandile_apur est un data.frame constitué des ménages de grandile ayant répondu à l'enquête fictive de grandile.
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
grandile_apur
}
\description{
Le fichier grandile a été constitué par des formateurs du Cefil. C'est une base
de données pédagogique utilisé dans beaucoup d'exercices tout au long de leur parcours.
}
\author{
Berangere \email{berangere.duplouy@insee.fr}
}
\keyword{data}
