% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_grandile_redress.R
\docType{data}
\name{grandile_redress}
\alias{grandile_redress}
\title{grandile_redress}
\format{
grandile_redress est un data.frame constitué des ménages de grandile dont les réponses ont été redressées.
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
grandile_redress
}
\description{
Le fichier grandile a été constitué par des formateurs du Cefil. C'est une base
de données pédagogique utilisé dans beaucoup d'exercices tout au long de leur parcours.
}
\author{
Berangere \email{berangere.duplouy@insee.fr}
}
\keyword{data}
