% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graph_month.R
\name{graph_month}
\alias{graph_month}
\title{graph_month}
\usage{
graph_month(serie_ts)
}
\arguments{
\item{serie_ts}{une série temporelle}
}
\value{
un graphique au format plot
}
\description{
Graphique pour étudier les séries avec saisonnalité
}
\examples{
graph_month(airpass)
}
