% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graphggplotly_diagbarre.R
\name{graphggplotly_diagbarre}
\alias{graphggplotly_diagbarre}
\title{graphggplotly_diagbarre}
\usage{
graphggplotly_diagbarre(data, var)
}
\arguments{
\item{data}{un dataframe}

\item{var}{une variable qualitative}
}
\value{
un plotly
}
\description{
Fonction qui permet de générer un diagrame en barre
avec une légende
}
\examples{
graphggplotly_diagbarre(grandile,var = "PCS")
}
