% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graphggplotly_nuage.R
\name{graphggplotly_nuage}
\alias{graphggplotly_nuage}
\title{graphggplotly_nuage}
\usage{
graphggplotly_nuage(data, var1, var2)
}
\arguments{
\item{data}{un dataframe}

\item{var1}{une variable quantitative}

\item{var2}{une variable quantitative}
}
\value{
un graphique ggplotly
}
\description{
Fonction pour produire un nuage de points avec la droite de  
regression linéaire associée
}
\examples{
graphggplotly_nuage(grandile,"REV_DISPONIBLE","PATRIMOINE")
}
