% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graphggplotly_qualiquanti.R
\name{graphggplotly_qualiquanti}
\alias{graphggplotly_qualiquanti}
\title{graphggplotly_qualiquanti}
\usage{
graphggplotly_qualiquanti(data, varquali, varquanti)
}
\arguments{
\item{data}{un dataframe}

\item{varquali}{une variable qualitative}

\item{varquanti}{une variable quantitative}
}
\value{
un graphique au format ggplot
}
\description{
fonction qui crée un nuage de points entre une qualitative
et une quantitative
}
\examples{
graphggplotly_qualiquanti(grandile,"PCS","REV_DISPONIBLE")
}
