% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_histoplotly_simul.R
\name{histoplotly_simul}
\alias{histoplotly_simul}
\title{histoplotly_simul}
\usage{
histoplotly_simul(serie, var)
}
\arguments{
\item{serie}{un vecteur numérique}

\item{var}{une variable quantitative}
}
\value{
un graphique GGPLOT2
}
\description{
Histogramme de la simulation d'estimation avec la moyenne empirique
}
\examples{
a <- simul_moyenne(grandile,"AGE",1000,1000)
histoplotly_simul(a,"AGE")
}
