% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_hotel.R
\docType{data}
\name{hotel}
\alias{hotel}
\title{hotel}
\format{
hotel est une série temporelle.
}
\source{
\url{https://www.insee.fr/fr/statistiques/serie/010599344#Telechargement} {Site de l'Insee}
}
\usage{
hotel
}
\description{
Le fichier hotel est la fréquentation hotelière en France métro
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
