% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_lineaireM_resid.R
\name{model_lineaireM_resid}
\alias{model_lineaireM_resid}
\title{model_lineaireM_resid}
\usage{
model_lineaireM_resid(
  input_data,
  var_expliquee,
  var_explicatives,
  constante = TRUE
)
}
\arguments{
\item{input_data}{un dataframe}

\item{var_expliquee}{une variable quantitative}

\item{var_explicatives}{un vecteur de variables quantitatives}

\item{constante}{un boolean}
}
\value{
un histogramme avec les résidus de la régression linéaire
}
\description{
Fonction qui fait tourner une regression lineaire multiple et sort un histogramme des residus
}
\examples{
model_lineaireM_resid(iris, "Petal.Length", c("Sepal.Length", "Petal.Width", "Sepal.Width"))
}
