% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_lineaireM_tab.R
\name{model_lineaireM_tab}
\alias{model_lineaireM_tab}
\title{model_lineaireM_tab}
\usage{
model_lineaireM_tab(
  input_data,
  var_expliquee,
  var_explicatives,
  constante = TRUE
)
}
\arguments{
\item{input_data}{un dataframe}

\item{var_expliquee}{une variable quantitative}

\item{var_explicatives}{un vecteur de variables quantitatives}
}
\value{
un tableau de resultat de la fontion lm()
}
\description{
Fonction qui fait tourner une regression lineaire multiple et sort un tableau de resultat
}
\examples{
model_lineaireM_tab(iris, "Petal.Length", c("Sepal.Length", "Petal.Width", "Sepal.Width"))
}
