% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_lineaireS_plot.R
\name{model_lineaireS_plot}
\alias{model_lineaireS_plot}
\title{model_lineaireS_plot}
\usage{
model_lineaireS_plot(input_data, var_expliquee, var_explicative, constante)
}
\arguments{
\item{input_data}{un dataframe}

\item{var_expliquee}{une variable quantitative}

\item{var_explicative}{une variable quantitative}

\item{constante}{un boolean}
}
\value{
un graphique ggplot
}
\description{
FOnction qui trace le nuage de point et la droite de regression lineaire
}
\examples{
model_lineaireS_plot(grandile, "REV_DISPONIBLE", "PATRIMOINE", TRUE)

}
