% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_lineaireS_tab.R
\name{model_lineaireS_tab}
\alias{model_lineaireS_tab}
\title{model_lineaireS_tab}
\usage{
model_lineaireS_tab(input_data, var_expliquee, var_explicative, constante)
}
\arguments{
\item{input_data}{un dataframe}

\item{var_expliquee}{variable quantitative}

\item{var_explicative}{variable quantitative}

\item{constante}{boolean}
}
\value{
un tableau de resultat de la fontion lm()
}
\description{
Fonction qui fait tourner 
une regression lineaire simple et sort un tableau de resultat
}
\examples{
model_lineaireS_tab(grandile, "REV_DISPONIBLE", "PATRIMOINE", TRUE)
}
