% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_logistique_tab.R
\name{model_logistique_tab}
\alias{model_logistique_tab}
\title{model_logistique_tab}
\usage{
model_logistique_tab(
  input_data,
  var_expliquee,
  var_explicatives,
  constante = TRUE
)
}
\arguments{
\item{input_data}{un dataframe}

\item{var_expliquee}{une variable quantitative}

\item{var_explicatives}{un vecteur de variables quantitatives}

\item{constante}{un boolean}
}
\value{
un tableau de resultat de la fontion glm()
}
\description{
Fonction qui fait tourner une regression logistique et sort un tableau de resultat
}
