% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_logistiqueSS_tab.R
\name{model_logistiqueSS_tab}
\alias{model_logistiqueSS_tab}
\title{model_logistiqueSS_tab}
\usage{
model_logistiqueSS_tab(
  input_data,
  var_expliquee,
  var_explicatives,
  constante = TRUE
)
}
\arguments{
\item{input_data}{un dataframe}

\item{var_expliquee}{une variable quantitative}

\item{var_explicatives}{un vecteur de variables quantitatives}

\item{constante}{un boolean}
}
\value{
un model glm()
}
\description{
Fonction qui fait tourner une regression logistique
}
\examples{
model_logistiqueSS_tab(input_data = titanic, 
var_expliquee = "Survived", 
var_explicatives = c("Sex","Age"), constante = TRUE)
}
