% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ozone.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{ozone}
\format{
ozone est un data.frame constitué de 1366 observations (rows) et de 13 variables (col).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
ozone
}
\description{
Le fichier ozone est constitué de 1366 relevés journaliers effectués à Rennes entre 1994 et 2001, 
pour mesurer la polution de l'air.
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
