% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_apur_m2.R
\name{plot_apur_m2}
\alias{plot_apur_m2}
\title{plot_apur_m2}
\usage{
plot_apur_m2(input_data, var_apur)
}
\arguments{
\item{input_data}{un dataframe}

\item{var_apur}{une variable quantitative}
}
\value{
a graph
}
\description{
A fct function that builds a plot with geompoints
}
