% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_ponderee.R
\name{pond_m2}
\alias{pond_m2}
\title{table_ponderee}
\usage{
pond_m2(data_pond, var_pond)
}
\arguments{
\item{data_pond}{un dataframe}

\item{var_pond}{une variable de redressement}
}
\value{
un dataframe
}
\description{
Fonction qui crée une table pondérée
}
\examples{
pond_m2(grandile_redress, "ACT")

}
