% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rambarplot_quali.R
\name{rambarplot_quali}
\alias{rambarplot_quali}
\title{rambarplot_quali}
\usage{
rambarplot_quali(data, var)
}
\arguments{
\item{data}{un dataframe}

\item{var}{variable qualitative}
}
\value{
un graphique ramcharts
}
\description{
Fonction qui crée un diagramme en barre de type rAmCharts
à partir d'une variable qualitative
}
\examples{
rambarplot_quali(grandile,"LIB_PCS")
}
