% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tableau_croise.R
\name{tableau_croise}
\alias{tableau_croise}
\title{tableau_croise}
\usage{
tableau_croise(data, varquali1, varquali2, ligne = FALSE)
}
\arguments{
\item{data}{un dataframe}

\item{varquali1}{une variable qualitative}

\item{varquali2}{une variable qualitative}

\item{ligne}{un bolean}
}
\value{
un tible
}
\description{
Fonction qui permet de produire un tableau croisé en profil ligne et colonne
}
\examples{
tableau_croise(grandile,"DIPL","PCS",ligne=TRUE)
}
