% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tirage_grappe_m2.R
\name{tirage_grappe_m2}
\alias{tirage_grappe_m2}
\title{Tirage_grappe_m2}
\usage{
tirage_grappe_m2(input_data, taille_grappe, var_grappe)
}
\arguments{
\item{input_data}{un dataframe}

\item{taille_grappe}{la taille de la grappe}

\item{var_grappe}{une variable de stratification}
}
\value{
un echantillon
}
\description{
Une fonction de tirage en grappe
}
\examples{
tirage_grappe_m2(grandile, 4, "DIPL")
}
