% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tirage_sas_m2.R
\name{tirage_sas_m2}
\alias{tirage_sas_m2}
\title{tirage_sas_m2}
\usage{
tirage_sas_m2(input_data, taille_sas)
}
\arguments{
\item{input_data}{un dataframe}

\item{taille_sas}{un nombre d'unités échantillonné}
}
\value{
a dataframe, result of a simple random sampling
}
\description{
A fct function that does a simple random sampling (sas)
}
