% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_vins.R
\docType{data}
\name{vins}
\alias{vins}
\title{vins}
\format{
vins est un data.frame constitué de 4 898 vins (rows) et de 12 variables (col).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
vins
}
\description{
Le fichier vins est constitué de 4898 individus qui sont des vins blancs portuguais de la 
région du Vinho Verde, avec leur composition chimique.
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
