% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_visiteurs.R
\docType{data}
\name{visiteurs}
\alias{visiteurs}
\title{visiteurs}
\format{
visiteurs est une série temporelle.
}
\source{
{Openclassroom}
}
\usage{
visiteurs
}
\description{
Le fichier visiteur est nombre de visiteurs d'un musée
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
