% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tirage_degres_m2.R
\name{tirage_degres_m2}
\alias{tirage_degres_m2}
\title{tirage_degres_m2}
\usage{
tirage_degres_m2(input_data, taille_UP, taille_tot, var_degres)
}
\arguments{
\item{input_data}{un dataframe}

\item{taille_UP}{un nombre d'unités primaires}

\item{taille_tot}{un nombre d'unités échantillonnées total}

\item{var_degres}{une variable de stratification}
}
\value{
a dataframe, result of a cluster sampling with 2 samplings
}
\description{
A fct function that does a cluster sampling with 2 sample random sampling
}
