% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tirage_grappe_m2.R
\name{tirage_grappe_m2}
\alias{tirage_grappe_m2}
\title{Tirage_grappe_m2}
\usage{
tirage_grappe_m2(input_data, taille_grappe, var_grappe)
}
\arguments{
\item{input_data}{}

\item{taille_grappe}{}

\item{var_grappe}{}
}
\value{
un echantillon
}
\description{
Tirage_grappe_m2
}
\examples{
tirage_grappe_m2(grandile, 4, "DIPL")
}
